xdp_samples: Flow tracker combo
This is a an example of using the same XDP2 parser compiled in both
a regular program and an XDP program. The regular program to parse
TCP and UDP packet over IPv4 and IPv6 and then print tuple insformation.
The XDP prggram uses the same parser to identify and track TCP flows

Signed-off-by: Tom Herbert <tom@herbertland.com>