bitmap: Multi-word bitmaps
Various operations for muli-word bitmaps.

bitmap.h is the frontend header file, bitmap_helpers.h is the
backend with most of the implementation of operations.

The header files contain functions and structures for processing
bitmaps. A bitmap is defined by a word size __u{8,16,32,64} * and a
number of bits (nbits). A bitmap may span multiple words and the bits
are numbered start with the low order bit of the first word. If N is
the number of bits in each word, the numbering is like:

	bit_0 = addr[0] & (1 << 0)
	bit_1 = addr[0] & (1 << 1)
	bit_2 = addr[0] & (1 << 2)
	...
	bit_(N-2) = addr[0] & (1 << (N-2))
	bit_(N-1) = addr[0] & (1 << (N-1))
	...
	bit_(nbits - 1) = addr[(nbits - 1) / N] & (1 << (nbits - 1) % N)

Reverse bitmaps structure the words in increasing order. If NUM_WORDS
is the maximum number of words in a bitmap then the bits are numbered
as:

	bit_0 = addr[NUM_WORDS - 1] & (1 << 0)
	bit_1 = addr[NUM_WORDS - 1] & (1 << 1)
	bit_2 = addr[NUM_WORDS - 1] & (1 << 2)
	...

The words of a __u16,  __u32, __u16 may be in little endian or big
endian byte order (also called "Network Byte Order"). Operations on
bitmaps that don't match the endinanness of the host can be done
using the "swap" variants of bitmap operations

Bitmap operation functions have the general format:

	xdp2_bitmap[8, 16, 32, 64][swp]_<operation>](bitmap, ARGS, ...)
	xdp2_rbitmap[8, 16, 32, 64][swp]_<operation>](bitmap, ARGS, ...,
						       NUM_WORDS)

If the 'r' is present before the bitmap then the bitmap is treated as
a reverse bitmap. A number (8, 16, 32, or 64) following bitmap indicates
the word size. If the number is absent the default size is used. "swp"
indicates that the bytes of a word are swapped for endinanness.
<operation> gives the bitmap operation, 'bitmap' referes to the input
bitmap as a pointer to the first words,  and the ARGS list gives the
arguments for the operation

If 'swp' is present then the 16, 32, or 64 bit words of a bitmap are
swapped relative to host byte order

If 'wsrc' is presenn then the first source operand is single word and
not a bit map. This acts as a pattern that can be applied to operations
on the second source

If destsrc is present then the desintation bitmap is also the (second)
source operand

If a number of bits is not specified in an operation then the type
is unsigned long. This matches the bitmaps used in the Linux kernel.
For these default types swap and reverse bitmaps are not defined

Signed-off-by: Tom Herbert <tom@herbertland.com>