vstruct: Add variable structures
Variable structures are a means to have multiple variable length arrays
in the same structure. Typically, at most one variable length array
can be used be declaring the last element is structure as [] in C99
flexible arrays syntax, like TYPE array[] (i.e. structure->array[2]
can be accessed presuming that a malloc of the structure is performed
with sufficient size.

Variable structures work by create a "variable structure map" or a
vsmap. A vsmap contains an array of offsets of each of the variable
length fields realtive to the data structure being defined, ie. variable
length arrays, in the structure. A vsmap for a structure is built as
runtime before the structure is allocated and used.

A variable structure is composed of a static region followed by a
variable region. The variable region is made up of two or more variable
length arrays of some element type. The sizes of the variable arrays
are only known at run time so these cannot be statically defined in the
structure

As an example, consider the following variable structure:

    struct my_struct {
    // Statically defined section
          struct my_struct_vsmap vsmap;
          unsigned int refcnt;
          struct socket *socket;
    // Variable section -- note that these are commented out
/*
	  struct my_work_item work_items[num_work_items];
	  struct my_object objects[num_work_items];
	  struct my_locks[num_locks];
*/
    }

The static region of struct my_struct defines three fields:
my_struct_vsmap, refcnt, and socket. These will be compiled normally be
the compiler as declared structure fields. The variable portion consists
of there fields: work_items, objects, and locks. The number of elements
for each of these is taken from runtime configuration.

Note that the my_struct_vsmap is a special data structure that describes
the offset of a variable structure. Each variable structure has an
associated vsmap that is used to access the variable arrays of the
structure. In the above example, the vsmap is include in the user's
structure definition, however it may be a separate data structure

The struct vsmap structure, e.g. struct my_struct_vsmap, is defined
using the XDP2_VSTRUCT macro. The first argument is the structure type
the vsmap is associate with. Following the structure type is a variable
list of pairs for the variable array where a pair consists of the field
name names and the type of an array element. For example, the vsmap for
my_struct example above would be at compiler time created by:

   XDP2_VSTRUCT(my_struct,
                 (work_items, struct my_work_item),
                 (objects, struct my_object),
                 (locks, struct my_locks)
  )

When this macro is called the output is the definition of a vsmap for
my_struct and is called my_struct_vsmap. This will be defined as:

   struct my_struct_vsmap {
          __u32 work_items_offset;
          __u32 objects_offset;
          __u32 locks_offset;
   }

Additionally, struct my_struct_def_vsmap is defined that can contain
additional information about fields-- this is only used in making a
vsmap or for debugging

When the program runs, the lengths for the arrays of the variable
structure are known and the vsmap can be instantiated. This is done by
the *XDP2_VSTRUCT_INSTANTIATE* macros. To instantiate the vsmap for a
my_struct, one might do:

   struct my_struct_vsmap vsmap; // vsmap embedded in the structure

   size = XDP2_VSTRUCT_INSTANTIATE(my_struct,
        (work_items, num_work_items, 0),
        (objects, num_work_items, 0),
        (locks, num_locks, XDP2_CACHE_LINE_SIZE)
   )

The effect of this macro will be to set the offsets in my_struct.vsmap.
For instance, if sizeof(struct my_work_item) is 112,
sizeof(struct my_object) is 60, and sizeof(struct my_locks) is 16,
num_work_items is 100, and num_locks is 1024, and XDP2_CACHE_LINE_SIZE
is sixty-four then after invoking the macro:

    work_items_offset = 128 // 112 aligned to cache line
    objects_offset =  6144 // 128+(100*60) = 6128, 6144 aligned
    work_items_offset = 12160    // 6144+(100*60) = 12144, 12160 aligned

    and return size = 17644

The size returned is that size for the variable structure. A working
instance can now be malloc'ed, for instance by:

    struct my_struct *my_data = malloc(size);
    *my_data->vsmap = vsmap

my_data is now an allocated varaible structure with the prescribed
lengths

To access of variable array the XDP2_VSTRUCT_GETPTR macro is invoked.
This take a pointer to the alloced variable structure and the name of
the field for the variable array. For instance,

    struct my_work_item *witem =
                XDP2_VSTRUCT_GETPTR(my_data, work_items);

returns a pointer to the first consumer fifo structure in the
variable array for consumer fifos in the example

Signed-off-by: Tom Herbert <tom@herbertland.com>