dynamic_tables: Define dynamic tables
Dynamic tables are defined and populate at run tine. There are
three variants: plain, ternary, and lpm, Each table is defined to
have a key of some size as input. WHen the table lookup function
is called the key is compared to the table entries. For a plain
tables all bits in the input key need to match the table key. For
a ternary table, each entry defines a mask. When a lookup is done
the mask is applied to the input key and the result is compared to
the key in the entry (also masked). For a lpm (longest prefix match)
table, each entry contains a prefix length. When a lookup is done,
the entry with the longest prefix that  matches the input key is
returned.

When an entry is matched, an assoicated handler function is called.

See documentation for more information

Signed-off-by: Tom Herbert <tom@herbertland.com>