pvbufs: Implementation of pvbufs and pbufs
pbufs or "packet buffers" are data buffers allocated from object
allocators to hold packet data

pvbufs or "packet vector bufferers" are organized as an array of iovecs
where each iovec that contain a pointer to a pbuf or an embedded pvbuf

pbufs can be difference sizes up to 1 << 20 bytes and are allocated
from a set of object allocators for pbufs of different sizes. The
object sizes for the allocator range 64 to 1 << 20 and each power
of two value in between. At initialization the object allocators are
initialized with some number of objects, and the number of objects
configured for an object allocator may be zero

To allocate a pbuf, xdp2_pbuf_alloc is called with the desired size
as an agument. The size is rounded up the nearest power of and an
object allocator supporting that size is selected. A larger object
may be allocated if there are no available objects

To allocate a pvbuf, xdp2_pvbuf_alloc with an argument for the size
of the data in the pvbuf to be allocated. If the size is non-zero
then pbufs are allocated to hold the size of data and are placed in
the iovecs of the pvbuf. If the size is zero then an empty pvbuf
structure is returned. xdp2_pvbuf_alloc may allocate a larger pbuf
to hold data, or may allocate a set of smaller ones to reduce slack
The fact argument is value in the range of 0..65536 and is the numerator
with a 65536 denominator to use to determine whether a smaller
allocation should be used

An iovec is composed of a forty-four bit address and a twenty bit
length, both pbufs and pvbufs have a forty-four bit reprensations.
The address for a pbufs includes a tag identifying the original
allocator, and object index that identifies the allocated object
relative to the allocator, and the byte offset within the object.
Given a tag, the base length of a buffer is 1 << (6 + tag). Tag
value 15 has special meaning; for object allocation is has the
same characteristics as tag value 14, however when the tag is used
in an iovec and the iovec length is zero then the real length is
(1 << 20).

pvbuf.h contains the definitions and APIs for pbufs and pvbufs. pvbuf.c
contains the backend library functions

Add struct xdp2_pbuf_init_allocator which includes number of object,
optional base pointer, and an optional allocator. If the last two are
NULL in initialization they are allocated, else the values are used
to initialize the pvbuf allocator. Change pvbuf initialization functions
to take and argument to struct xdp2_pbuf_init_allocator

- Add booleans to the xdp2_pbuf allocator structure that indicate that
  the this pbuf allocator, the referenced allocator, or the base object
  memory for the allocator was malloced. These are used when the pbuf
  allocator is being freed to determine if the memory objects should be
  freed
- Add booleans to the xdp2_pvbuf_mgr structure that indicate that
  the references allocator or base object memory for the allocator was
  malloced. These are used when the pvbuf manager is being freed to
  determine if the memory objects should be freed
- Add an address translation number to a pvbuf manager. This is used
  to provide address translations of pointers in the pvbuf manager when
  it resides in shared memory
- Add a pointer to struct xdp2_pbuf_init_allocator to a
  xdp2_pbuf_allocator. This allows the pbuf allocator to use
  pre-allocated memory instead of always malloc'ing it
- Add helper function XDP2_PVBUF_GET_ALLOCATOR that applies address
  translation to get the real address for the pvbuf_allocator of a pvbuf
  manager
- Add helper function XDP2_PVBUF_GET_ADDRESS that applies address
  translation to get the real address for addresses in a pvbuf manager

- Added xdp2_pvbuf_get_tag_from_addr44
- Added copy data to pvbuf and copy pvbuf to data functions to move
  data to/from pvbufs from/to a linear buffer
- Added xdp2_pvbuf_checksum to compute the checksum over a pvbuf
- Added a comment providing more details on size limits and the
  fact argument in pvbuf allocation

In a xdp2_paddr_t rework the pvbuf address fields so that the
forty-four bit address is in the high order bits instead of the
low order bits. This make changing the address size easier (i.e.
changing to something other than a forty-four bit address)

- Move twenty bit unused field in struct xdp2_pvbuf_paddr to
  be low order bits
- Rename _val to paddr in struct xdp2_pvbuf_paddr - Make name of paddr
  and xdp2_pvbuf_add consitent (paddr is a xdp2_paddr_t value, pvaddr
  and pbaddr refer to struct xdp2_pvbuf_paddr and struct xdp2_pbuf_paddr
  addresses
- Switch length and address fields in struct xdp2_iovec
- Rename _val to paddr in struct xdp2_iovec
- Remove struct xdp2_iovec_paddr, absorb in struct xdp2_iovec
- Add xdp2_pvbuf_set_iovec and xdp2_pvbuf_zero_iovec helper
  macros
- Remove unnecessary __xdp2_pbuf_get_zindex
- Use iovec->paddr instead of iovec->address to handle shift of
  address to high order bits
- Simplify argument names for paddr's
- Rename find_last_iovec to xdp2_pvbuf_find_last_iovec
- Add information to XDP2_ASSERTs

A one-ref is one that has a single reference as indicated by the
buffer type in the address. When freeing a single-ref pbuf, the pbuf
can be freed immediately without decrementing its reference count
(e.g. by enqueuing on an object FIFO).

A pbuf can be allocated as a single-ref pbuf or regular pbuf, the
address contains a different buffer tag. If a single-ref pbuf is
cloned or requires an additional reference, it can be converted to
a regular pbuf. This is done by changing the address to indicate a
regular pbuf and the caller must *only* use the new address going
forward
ease enter the commit message for your changes. Lines starting
th '#' will be ignored, and an empty message aborts the commit.

More pvbuf

Signed-off-by: Tom Herbert <tom@herbertland.com>